+++ tests/framework/web/ViewTest.php
@@ -20,7 +20,7 @@ class ViewTest extends TestCase
     {
         parent::setUp();
     }
-    
+
     public function testRegisterJsVar()
     {
         $this->mockWebApplication([
@@ -36,12 +36,13 @@ class ViewTest extends TestCase
         $view->registerJsVar('username', 'samdark');
         $html = $view->render('@yiiunit/data/views/layout.php', ['content' => 'content']);
         $this->assertContains('<script>var username = "samdark";</script></head>', $html);
-        
+
         $view = new View();
-        $view->registerJsVar('objectTest', [
-            'number' => 42,
-            'question' => 'Unknown',
-        ]);
+        $view->registerJsVar('objectTest',
+            [
+                'number' => 42,
+                'question' => 'Unknown',
+            ]);
         $html = $view->render('@yiiunit/data/views/layout.php', ['content' => 'content']);
         $this->assertContains('<script>var objectTest = {"number":42,"question":"Unknown"};</script></head>', $html);
     }
@@ -71,6 +72,12 @@ class ViewTest extends TestCase
         $view->registerJsFile('@web/js/somefile.js', ['position' => View::POS_END]);
         $html = $view->render('@yiiunit/data/views/layout.php', ['content' => 'content']);
         $this->assertContains('<script src="/baseUrl/js/somefile.js"></script></body>', $html);
+
+        // alias with depends
+        $view = new View();
+        $view->registerJsFile('@web/js/somefile.js', ['position' => View::POS_END, 'depends' => 'yii\web\AssetBundle']);
+        $html = $view->render('@yiiunit/data/views/layout.php', ['content' => 'content']);
+        $this->assertContains('<script src="/baseUrl/js/somefile.js"></script></body>', $html);
     }
 
     public function testRegisterCssFileWithAlias()
@@ -86,6 +93,13 @@ class ViewTest extends TestCase
 
         $view = new View();
         $view->registerCssFile('@web/css/somefile.css');
+        $html = $view->render('@yiiunit/data/views/layout.php', ['content' => 'content']);
+        $this->assertContains('<link href="/baseUrl/css/somefile.css" rel="stylesheet"></head>', $html);
+
+        // with depends
+        $view = new View();
+        $view->registerCssFile('@web/css/somefile.css',
+            ['position' => View::POS_END, 'depends' => 'yii\web\AssetBundle']);
         $html = $view->render('@yiiunit/data/views/layout.php', ['content' => 'content']);
         $this->assertContains('<link href="/baseUrl/css/somefile.css" rel="stylesheet"></head>', $html);
     }
@@ -136,5 +150,305 @@ class ViewTest extends TestCase
         }
 
         return $matches[1];
+    }
+
+    private function setUpAliases()
+    {
+        \Yii::setAlias('@web', '/');
+        \Yii::setAlias('@webroot', '@yiiunit/data/web');
+        \Yii::setAlias('@testAssetsPath', '@webroot/assets');
+        \Yii::setAlias('@testAssetsUrl', '@web/assets');
+        \Yii::setAlias('@testSourcePath', '@webroot/assetSources');
+    }
+
+    public function testAppendTimestampForRegisterJsFile()
+    {
+        $this->mockWebApplication([
+            'components' => [
+                'request' => [
+                    'scriptFile' => __DIR__ . '/baseUrl/index.php',
+                    'scriptUrl' => '/baseUrl/index.php',
+                ],
+            ],
+        ]);
+
+        $this->setUpAliases();
+
+        $pattern = '/assetSources\/js\/jquery\.js\?v\=\d+"/';
+
+        \Yii::$app->assetManager->appendTimestamp = true;
+
+        // will be used AssetManager and timestamp
+        $view = new View();
+        $view->registerJsFile('/assetSources/js/jquery.js',
+            ['depends' => 'yii\web\AssetBundle']); // <script src="/assetSources/js/jquery.js?v=1541056962"></script>
+        $html = $view->render('@yiiunit/data/views/layout.php', ['content' => 'content']);
+        $this->assertRegExp($pattern, $html);
+
+        // won't be used AssetManager but the timestamp will be
+        $view = new View();
+        $view->registerJsFile('/assetSources/js/jquery.js'); // <script src="/assetSources/js/jquery.js?v=1541056962"></script>
+        $html = $view->render('@yiiunit/data/views/layout.php', ['content' => 'content']);
+        $this->assertRegExp($pattern, $html);
+
+        $view = new View();
+        $view->registerJsFile('/assetSources/js/jquery.js',
+            ['appendTimestamp' => true]); // <script src="/assetSources/js/jquery.js?v=1541056962"></script>
+        $html = $view->render('@yiiunit/data/views/layout.php', ['content' => 'content']);
+        $this->assertRegExp($pattern, $html);
+
+        // redefine AssetManager timestamp setting
+        $view = new View();
+        $view->registerJsFile('/assetSources/js/jquery.js',
+            ['appendTimestamp' => false]); // <script src="/assetSources/js/jquery.js"></script>
+        $html = $view->render('@yiiunit/data/views/layout.php', ['content' => 'content']);
+        $this->assertNotRegExp($pattern, $html);
+
+        // with alias
+        $view = new View();
+        $view->registerJsFile('@web/assetSources/js/jquery.js'); // <script src="/assetSources/js/jquery.js?v=1541056962"></script>
+        $html = $view->render('@yiiunit/data/views/layout.php', ['content' => 'content']);
+        $this->assertRegExp($pattern, $html);
+
+        // with alias but wo timestamp
+        // The timestamp setting won't be redefined because global AssetManager is used
+        $view = new View();
+        $view->registerJsFile('@web/assetSources/js/jquery.js',
+            [
+                'appendTimestamp' => false,
+                'depends' => 'yii\web\AssetBundle',
+            ]); // <script src="/assetSources/js/jquery.js"></script>
+        $html = $view->render('@yiiunit/data/views/layout.php', ['content' => 'content']);
+        $this->assertRegExp($pattern, $html);
+
+        // wo depends == wo AssetManager
+        $view = new View();
+        $view->registerJsFile('@web/assetSources/js/jquery.js',
+            ['appendTimestamp' => false]); // <script src="/assetSources/js/jquery.js"></script>
+        $html = $view->render('@yiiunit/data/views/layout.php', ['content' => 'content']);
+        $this->assertNotRegExp($pattern, $html);
+
+        // absolute link
+        $view = new View();
+        $view->registerJsFile('http://ajax.googleapis.com/ajax/libs/jquery/2.1.1/jquery.min.js');
+        $html = $view->render('@yiiunit/data/views/layout.php', ['content' => 'content']);
+        $this->assertContains('<script src="http://ajax.googleapis.com/ajax/libs/jquery/2.1.1/jquery.min.js"></script>',
+            $html);
+
+        $view = new View();
+        $view->registerJsFile('//ajax.googleapis.com/ajax/libs/jquery/2.1.1/jquery.min.js',
+            ['depends' => 'yii\web\AssetBundle']);
+        $html = $view->render('@yiiunit/data/views/layout.php', ['content' => 'content']);
+        $this->assertContains('<script src="//ajax.googleapis.com/ajax/libs/jquery/2.1.1/jquery.min.js"></script>',
+            $html);
+
+        $view = new View();
+        $view->registerJsFile('http://ajax.googleapis.com/ajax/libs/jquery/2.1.1/jquery.min.js',
+            ['depends' => 'yii\web\AssetBundle']);
+        $html = $view->render('@yiiunit/data/views/layout.php', ['content' => 'content']);
+        $this->assertContains('<script src="http://ajax.googleapis.com/ajax/libs/jquery/2.1.1/jquery.min.js"></script>',
+            $html);
+
+        \Yii::$app->assetManager->appendTimestamp = false;
+
+        $view = new View();
+        $view->registerJsFile('/assetSources/js/jquery.js',
+            ['depends' => 'yii\web\AssetBundle']); // <script src="/assetSources/js/jquery.js"></script>
+        $html = $view->render('@yiiunit/data/views/layout.php', ['content' => 'content']);
+        $this->assertNotRegExp($pattern, $html);
+
+        $view = new View();
+        $view->registerJsFile('/assetSources/js/jquery.js'); // <script src="/assetSources/js/jquery.js"></script>
+        $html = $view->render('@yiiunit/data/views/layout.php', ['content' => 'content']);
+        $this->assertNotRegExp($pattern, $html);
+
+        $view = new View();
+        $view->registerJsFile('/assetSources/js/jquery.js',
+            ['appendTimestamp' => true]); // <script src="/assetSources/js/jquery.js?v=1541056962"></script>
+        $html = $view->render('@yiiunit/data/views/layout.php', ['content' => 'content']);
+        $this->assertRegExp($pattern, $html);
+
+        // The timestamp setting won't be redefined because global AssetManager is used
+        $view = new View();
+        $view->registerJsFile('/assetSources/js/jquery.js',
+            [
+                'appendTimestamp' => true,
+                'depends' => 'yii\web\AssetBundle',
+            ]); // <script src="/assetSources/js/jquery.js"></script>
+        $html = $view->render('@yiiunit/data/views/layout.php', ['content' => 'content']);
+        $this->assertNotRegExp($pattern, $html);
+
+        $view = new View();
+        $view->registerJsFile('/assetSources/js/jquery.js',
+            ['appendTimestamp' => false]); // <script src="/assetSources/js/jquery.js"></script>
+        $html = $view->render('@yiiunit/data/views/layout.php', ['content' => 'content']);
+        $this->assertNotRegExp($pattern, $html);
+
+        // absolute link
+        $view = new View();
+        $view->registerJsFile('http://ajax.googleapis.com/ajax/libs/jquery/2.1.1/jquery.min.js');
+        $html = $view->render('@yiiunit/data/views/layout.php', ['content' => 'content']);
+        $this->assertContains('<script src="http://ajax.googleapis.com/ajax/libs/jquery/2.1.1/jquery.min.js"></script>',
+            $html);
+
+        $view = new View();
+        $view->registerJsFile('//ajax.googleapis.com/ajax/libs/jquery/2.1.1/jquery.min.js',
+            ['depends' => 'yii\web\AssetBundle']);
+        $html = $view->render('@yiiunit/data/views/layout.php', ['content' => 'content']);
+        $this->assertContains('<script src="//ajax.googleapis.com/ajax/libs/jquery/2.1.1/jquery.min.js"></script>',
+            $html);
+
+        $view = new View();
+        $view->registerJsFile('http://ajax.googleapis.com/ajax/libs/jquery/2.1.1/jquery.min.js',
+            ['depends' => 'yii\web\AssetBundle']);
+        $html = $view->render('@yiiunit/data/views/layout.php', ['content' => 'content']);
+        $this->assertContains('<script src="http://ajax.googleapis.com/ajax/libs/jquery/2.1.1/jquery.min.js"></script>',
+            $html);
+
+    }
+
+    public function testAppendTimestampForRegisterCssFile()
+    {
+        $this->mockWebApplication([
+            'components' => [
+                'request' => [
+                    'scriptFile' => __DIR__ . '/baseUrl/index.php',
+                    'scriptUrl' => '/baseUrl/index.php',
+                ],
+            ],
+        ]);
+
+        $this->setUpAliases();
+
+        $pattern = '/assetSources\/css\/stub\.css\?v\=\d+"/';
+
+        \Yii::$app->assetManager->appendTimestamp = true;
+
+        // will be used AssetManager and timestamp
+        $view = new View();
+        $view->registerCssFile('/assetSources/css/stub.css',
+            ['depends' => 'yii\web\AssetBundle']); // <link href="/assetSources/css/stub.css?v=1541056962" rel="stylesheet" >
+        $html = $view->render('@yiiunit/data/views/layout.php', ['content' => 'content']);
+        $this->assertRegExp($pattern, $html);
+
+        // won't be used AssetManager but the timestamp will be
+        $view = new View();
+        $view->registerCssFile('/assetSources/css/stub.css'); // <link href="/assetSources/css/stub.css?v=1541056962" rel="stylesheet" >
+        $html = $view->render('@yiiunit/data/views/layout.php', ['content' => 'content']);
+        $this->assertRegExp($pattern, $html);
+
+        $view = new View();
+        $view->registerCssFile('/assetSources/css/stub.css',
+            ['appendTimestamp' => true]); // <link href="/assetSources/css/stub.css?v=1541056962" rel="stylesheet" >
+        $html = $view->render('@yiiunit/data/views/layout.php', ['content' => 'content']);
+        $this->assertRegExp($pattern, $html);
+
+        // redefine AssetManager timestamp setting
+        $view = new View();
+        $view->registerCssFile('/assetSources/css/stub.css',
+            ['appendTimestamp' => false]); // <link href="/assetSources/css/stub.css" rel="stylesheet" >
+        $html = $view->render('@yiiunit/data/views/layout.php', ['content' => 'content']);
+        $this->assertNotRegExp($pattern, $html);
+
+        // with alias
+        $view = new View();
+        $view->registerCssFile('@web/assetSources/css/stub.css'); // <link href="/assetSources/css/stub.css?v=1541056962" rel="stylesheet" >
+        $html = $view->render('@yiiunit/data/views/layout.php', ['content' => 'content']);
+        $this->assertRegExp($pattern, $html);
+
+        // with alias but wo timestamp
+        // The timestamp setting won't be redefined because global AssetManager is used
+        $view = new View();
+        $view->registerCssFile('@web/assetSources/css/stub.css',
+            [
+                'appendTimestamp' => false,
+                'depends' => 'yii\web\AssetBundle',
+            ]); // <link href="/assetSources/css/stub.css" rel="stylesheet" >
+        $html = $view->render('@yiiunit/data/views/layout.php', ['content' => 'content']);
+        $this->assertRegExp($pattern, $html);
+
+        // wo depends == wo AssetManager
+        $view = new View();
+        $view->registerCssFile('@web/assetSources/css/stub.css',
+            ['appendTimestamp' => false]); // <link href="/assetSources/css/stub.css" rel="stylesheet" >
+        $html = $view->render('@yiiunit/data/views/layout.php', ['content' => 'content']);
+        $this->assertNotRegExp($pattern, $html);
+
+        // absolute link
+        $view = new View();
+        $view->registerCssFile('https://cdnjs.cloudflare.com/ajax/libs/balloon-css/1.0.3/balloon.css');
+        $html = $view->render('@yiiunit/data/views/layout.php', ['content' => 'content']);
+        $this->assertContains('<link href="https://cdnjs.cloudflare.com/ajax/libs/balloon-css/1.0.3/balloon.css" rel="stylesheet">',
+            $html);
+
+        $view = new View();
+        $view->registerCssFile('//cdnjs.cloudflare.com/ajax/libs/balloon-css/1.0.3/balloon.css',
+            ['depends' => 'yii\web\AssetBundle']);
+        $html = $view->render('@yiiunit/data/views/layout.php', ['content' => 'content']);
+        $this->assertContains('<link href="//cdnjs.cloudflare.com/ajax/libs/balloon-css/1.0.3/balloon.css" rel="stylesheet">',
+            $html);
+
+        $view = new View();
+        $view->registerCssFile('https://cdnjs.cloudflare.com/ajax/libs/balloon-css/1.0.3/balloon.css',
+            ['depends' => 'yii\web\AssetBundle']);
+        $html = $view->render('@yiiunit/data/views/layout.php', ['content' => 'content']);
+        $this->assertContains('<link href="https://cdnjs.cloudflare.com/ajax/libs/balloon-css/1.0.3/balloon.css" rel="stylesheet">',
+            $html);
+
+        \Yii::$app->assetManager->appendTimestamp = false;
+
+        $view = new View();
+        $view->registerCssFile('/assetSources/css/stub.css',
+            ['depends' => 'yii\web\AssetBundle']); // <link href="/assetSources/css/stub.css" rel="stylesheet" >
+        $html = $view->render('@yiiunit/data/views/layout.php', ['content' => 'content']);
+        $this->assertNotRegExp($pattern, $html);
+
+        $view = new View();
+        $view->registerCssFile('/assetSources/css/stub.css'); // <link href="/assetSources/css/stub.css" rel="stylesheet" >
+        $html = $view->render('@yiiunit/data/views/layout.php', ['content' => 'content']);
+        $this->assertNotRegExp($pattern, $html);
+
+        $view = new View();
+        $view->registerCssFile('/assetSources/css/stub.css',
+            ['appendTimestamp' => true]); // <link href="/assetSources/css/stub.css?v=1541056962" rel="stylesheet" >
+        $html = $view->render('@yiiunit/data/views/layout.php', ['content' => 'content']);
+        $this->assertRegExp($pattern, $html);
+
+        // The timestamp setting won't be redefined because global AssetManager is used
+        $view = new View();
+        $view->registerCssFile('/assetSources/css/stub.css',
+            [
+                'appendTimestamp' => true,
+                'depends' => 'yii\web\AssetBundle',
+            ]); // <link href="/assetSources/css/stub.css" rel="stylesheet" >
+        $html = $view->render('@yiiunit/data/views/layout.php', ['content' => 'content']);
+        $this->assertNotRegExp($pattern, $html);
+
+        $view = new View();
+        $view->registerCssFile('/assetSources/css/stub.css',
+            ['appendTimestamp' => false]); // <link href="/assetSources/css/stub.css" rel="stylesheet" >
+        $html = $view->render('@yiiunit/data/views/layout.php', ['content' => 'content']);
+        $this->assertNotRegExp($pattern, $html);
+
+        // absolute link
+        $view = new View();
+        $view->registerCssFile('https://cdnjs.cloudflare.com/ajax/libs/balloon-css/1.0.3/balloon.css');
+        $html = $view->render('@yiiunit/data/views/layout.php', ['content' => 'content']);
+        $this->assertContains('<link href="https://cdnjs.cloudflare.com/ajax/libs/balloon-css/1.0.3/balloon.css" rel="stylesheet">',
+            $html);
+
+        $view = new View();
+        $view->registerCssFile('//cdnjs.cloudflare.com/ajax/libs/balloon-css/1.0.3/balloon.css',
+            ['depends' => 'yii\web\AssetBundle']);
+        $html = $view->render('@yiiunit/data/views/layout.php', ['content' => 'content']);
+        $this->assertContains('<link href="//cdnjs.cloudflare.com/ajax/libs/balloon-css/1.0.3/balloon.css" rel="stylesheet">',
+            $html);
+
+        $view = new View();
+        $view->registerCssFile('https://cdnjs.cloudflare.com/ajax/libs/balloon-css/1.0.3/balloon.css',
+            ['depends' => 'yii\web\AssetBundle']);
+        $html = $view->render('@yiiunit/data/views/layout.php', ['content' => 'content']);
+        $this->assertContains('<link href="https://cdnjs.cloudflare.com/ajax/libs/balloon-css/1.0.3/balloon.css" rel="stylesheet">',
+            $html);
     }
 }